+++ tests/testthat/test-normality_tests.R
@@ -1,4 +1,7 @@
 test_that("normality tests", {
-  res = normality_tests(iris)
-  expect_type(res, "list")
+  suppressWarnings({
+    res = normality_tests(iris)
+    expect_type(res, "list")
+  })
+
 })