+++ tests/testthat/test-univar_norm_tests.R
@@ -1,4 +1,6 @@
 test_that("univar norm test", {
-  res = univar_norm_tests(iris[,1],colnames(iris)[1])
-  expect_type(res,'double')
+  suppressWarnings({
+    res = univar_norm_tests(iris[,1],colnames(iris)[1])
+    expect_type(res,'double')
+  })
 })